Remove duplicate print_run_time, and record the time of image search 去除重复的print_run_time，并且记录图像查找的时间
(cherry picked from commit 2990450a29893730403be5751ba9e85e298203c4)
